Splitted configuration, added todos, minor changes
