+++ hermes-consumers/src/main/java/pl/allegro/tech/hermes/consumers/di/config/HealthConfiguration.java
@@ -4,13 +4,10 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import pl.allegro.tech.hermes.common.config.ConfigFactory;
-import pl.allegro.tech.hermes.common.config.Configs;
 import pl.allegro.tech.hermes.consumers.health.ConsumerHttpServer;
 import pl.allegro.tech.hermes.consumers.health.ConsumerMonitor;
 
 import java.io.IOException;
-
-import static pl.allegro.tech.hermes.common.config.Configs.KAFKA_CONSUMER_HEARTBEAT_INTERVAL_MS_CONFIG;
 
 @Configuration
 public class HealthConfiguration {
@@ -19,10 +16,6 @@ public class HealthConfiguration {
     public ConsumerHttpServer consumerHttpServer(ConfigFactory configFactory,
                                                  ConsumerMonitor monitor,
                                                  ObjectMapper mapper) throws IOException {
-//        Configs.CONSUMER_HEALTH_CHECK_PORT 7454
-//        Configs.METRICS_GRAPHITE_REPORTER false
-//        Configs.METRICS_ZOOKEEPER_REPORTER false
-//        overrideProperty(KAFKA_CONSUMER_HEARTBEAT_INTERVAL_MS_CONFIG, 50);
         return new ConsumerHttpServer(configFactory, monitor, mapper);
     }
