+++ hermes-consumers/src/main/java/pl/allegro/tech/hermes/consumers/di/config/HermesConsumersConfiguration.java
@@ -1,6 +1,5 @@
 package pl.allegro.tech.hermes.consumers.di.config;
 
-import org.springframework.context.ApplicationContext;
 import org.springframework.context.ConfigurableApplicationContext;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
