+++ integration/src/integration/java/pl/allegro/tech/hermes/integration/management/OfflineRetransmissionManagementTest.java
@@ -65,7 +65,7 @@ public class OfflineRetransmissionManagementTest extends IntegrationTest {
     }
 
     @Test
-    public void shouldReturnClientErrorWhenRequestingRetransmissionWithEmptyData() {
+    public void shouldReturnClientErrorWhenRequestingRetransmissionWithEmptyData() {//TODO: fix?
         // when
         OfflineRetransmissionRequest request = new OfflineRetransmissionRequest(
                 "",
